code styling move utils to separate folder
