+++ ui/src/App.tsx
@@ -27,7 +27,6 @@ import {
   generateUrlHashWithFlowId,
   openWorkflowInNewTab,
   validateOrSaveAllJsonFileMyWorkflows,
-  updateRecentlyOpenedFileList,
 } from "./utils";
 import { Topbar } from "./topbar/Topbar";
 import { Workflow, WorkflowVersion } from "./types/dbTypes";
@@ -46,6 +45,7 @@ import { useStateRef } from "./customHooks/useStateRef";
 import { indexdb } from "./db-tables/indexdb";
 import EnableTwowaySyncConfirm from "./settings/EnableTwowaySyncConfirm";
 import { deepJsonDiffCheck } from "./utils/deepJsonDiffCheck";
+import { updateRecentlyOpenedFileList } from "./utils/recentOpenedFilesUtils";
 const AppIsDirtyEventListener = React.lazy(
   () => import("./topbar/AppIsDirtyEventListener"),
 );
@@ -193,7 +193,7 @@ export default function App() {
     if (latestWfID) {
       loadWorkflowIDImpl(latestWfID);
     }
-    await validateOrSaveAllJsonFileMyWorkflows();
+    // await validateOrSaveAllJsonFileMyWorkflows();
     const twoway = await userSettingsTable?.getSetting("twoWaySync");
     !twoway &&
       indexdb.cache.get(UPGRADE_TO_2WAY_SYNC_KEY).then(async (value) => {
