+++ ui/src/utils.tsx
@@ -5,20 +5,13 @@ import {
   generateFilePathAbsolute,
   saveJsonFileMyWorkflows,
 } from "./db-tables/DiskFileUtils";
-import {
-  Folder,
-  Workflow,
-  EShortcutKeys,
-  RecentlyOpenedFile,
-} from "./types/dbTypes";
+import { Folder, Workflow, EShortcutKeys } from "./types/dbTypes";
 // @ts-expect-error ComfyUI import
 import { app } from "/scripts/app.js";
 import {
   COMFYSPACE_TRACKING_FIELD_NAME,
   LEGACY_COMFYSPACE_TRACKING_FIELD_NAME,
-  RECENTLY_OPENED_FILE_LIST,
 } from "./const";
-import { indexdb } from "./db-tables/indexdb";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 declare let LiteGraph: any, LGraph: any, LGraphCanvas: any;
@@ -350,36 +343,4 @@ export async function rewriteAllLocalFiles() {
       console.error(error);
     }
   }
-}
-
-export async function getRecentlyOpenedFileList() {
-  const fileListCache = await indexdb.cache.get(RECENTLY_OPENED_FILE_LIST);
-  let recentlyOpenedFileList: RecentlyOpenedFile[] = [];
-  try {
-    fileListCache?.value &&
-      (recentlyOpenedFileList = JSON.parse(fileListCache?.value));
-  } catch (e) {
-    console.error("error parsing recently opened file list", e);
-  }
-  return recentlyOpenedFileList;
-}
-
-export async function updateRecentlyOpenedFileList(
-  curOpenFile: RecentlyOpenedFile,
-) {
-  const recentlyOpenedFileList = await getRecentlyOpenedFileList();
-  const findIndex = recentlyOpenedFileList.findIndex(
-    (file) => file.id === curOpenFile.id,
-  );
-  if (findIndex >= 0) {
-    recentlyOpenedFileList.splice(findIndex, 1);
-  }
-  recentlyOpenedFileList.unshift(curOpenFile);
-  if (recentlyOpenedFileList.length >= 20) {
-    recentlyOpenedFileList.length = 20;
-  }
-  indexdb.cache.put({
-    id: RECENTLY_OPENED_FILE_LIST,
-    value: JSON.stringify(recentlyOpenedFileList),
-  });
 }
